add screen frame children sorting

+++ src/main/kotlin/data/nodes/properties/Rectangle.kt
@@ -1,8 +1,8 @@
 package data.nodes.properties
 
 data class Rectangle(
-    val x: Int,
-    val y: Int,
-    val width: Int,
-    val height: Int
+    val x: Double,
+    val y: Double,
+    val width: Double,
+    val height: Double
 )
